backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pymupdf
sentence-transformers
faiss-cpu
openai